+++ modules/core/src/main/kotlin/com/github/l34130/mise/core/execution/configuration/MiseTomlTaskRunConfigurationProducer.kt
@@ -1,12 +1,15 @@
 package com.github.l34130.mise.core.execution.configuration
 
 import com.github.l34130.mise.core.lang.psi.MiseTomlFile
+import com.github.l34130.mise.core.lang.psi.taskName
 import com.intellij.execution.actions.ConfigurationContext
 import com.intellij.execution.actions.LazyRunConfigurationProducer
 import com.intellij.execution.configurations.ConfigurationFactory
 import com.intellij.openapi.components.PathMacroManager
 import com.intellij.openapi.util.Ref
 import com.intellij.psi.PsiElement
+import com.intellij.psi.util.parentOfType
+import org.toml.lang.psi.TomlTable
 
 class MiseTomlTaskRunConfigurationProducer : LazyRunConfigurationProducer<MiseTomlTaskRunConfiguration>() {
     override fun getConfigurationFactory(): ConfigurationFactory =
@@ -15,13 +18,7 @@ class MiseTomlTaskRunConfigurationProducer : LazyRunConfigurationProducer<MiseTo
     override fun isConfigurationFromContext(
         configuration: MiseTomlTaskRunConfiguration,
         context: ConfigurationContext,
-    ): Boolean {
-        val macroManager = PathMacroManager.getInstance(context.project)
-//        return macroManager.expandPath(configuration.filename) == context.location?.virtualFile?.path &&
-//                configuration.target == findTarget(context)?.name
-
-        return true
-    }
+    ): Boolean = configuration.miseTaskName == findTaskName(context)
 
     override fun setupConfigurationFromContext(
         configuration: MiseTomlTaskRunConfiguration,
@@ -30,11 +27,21 @@ class MiseTomlTaskRunConfigurationProducer : LazyRunConfigurationProducer<MiseTo
     ): Boolean {
         if (context.psiLocation?.containingFile !is MiseTomlFile) return false
 
-        val macroManager = PathMacroManager.getInstance(context.project)
-        val path = context.location?.virtualFile?.path
+        val taskName = findTaskName(context) ?: return false
+        configuration.miseTaskName = taskName
 
-        // edit configuration
+        val macroManager = PathMacroManager.getInstance(context.project)
+
+        val virtualFile = context.location?.virtualFile
+        configuration.workingDirectory = macroManager.collapsePath(virtualFile?.parent?.path ?: context.project.basePath)
+
+        configuration.name = "Run $taskName"
 
         return true
+    }
+
+    private fun findTaskName(context: ConfigurationContext): String? {
+        val table = context.psiLocation?.parentOfType<TomlTable>() ?: return null
+        return table.taskName
     }
 }