+++ modules/core/src/main/kotlin/com/github/l34130/mise/core/lang/MiseTomlFileType.kt
@@ -2,9 +2,10 @@ package com.github.l34130.mise.core.lang
 
 import com.github.l34130.mise.core.icon.MiseIcons
 import com.intellij.openapi.fileTypes.LanguageFileType
+import org.toml.lang.TomlLanguage
 import javax.swing.Icon
 
-object MiseTomlFileType : LanguageFileType(MiseTomlLanguage) {
+object MiseTomlFileType : LanguageFileType(TomlLanguage) {
     override fun getName(): String = "mise"
 
     override fun getDescription(): String = "Mise Configuration file"