+++ modules/core/src/main/kotlin/com/github/l34130/mise/core/lang/psi/MiseTomlFile.kt
@@ -1,13 +1,13 @@
 package com.github.l34130.mise.core.lang.psi
 
 import com.github.l34130.mise.core.lang.MiseTomlFileType
-import com.github.l34130.mise.core.lang.MiseTomlLanguage
 import com.intellij.extapi.psi.PsiFileBase
 import com.intellij.psi.FileViewProvider
+import org.toml.lang.TomlLanguage
 
 class MiseTomlFile(
     viewProvider: FileViewProvider,
-) : PsiFileBase(viewProvider, MiseTomlLanguage) {
+) : PsiFileBase(viewProvider, TomlLanguage) {
     override fun getFileType() = MiseTomlFileType
 
     override fun toString() = "Mise Toml File"