+++ modules/core/src/main/kotlin/com/github/l34130/mise/core/lang/psi/MiseTomlPsiPatterns.kt
@@ -1,11 +1,11 @@
 package com.github.l34130.mise.core.lang.psi
 
 import com.github.l34130.mise.core.lang.MiseTomlFileType
+import com.intellij.patterns.ElementPattern
 import com.intellij.patterns.ObjectPattern
 import com.intellij.patterns.PatternCondition
 import com.intellij.patterns.PlatformPatterns
 import com.intellij.patterns.PsiElementPattern
-import com.intellij.patterns.StandardPatterns
 import com.intellij.patterns.VirtualFilePattern
 import com.intellij.psi.PsiElement
 import com.intellij.util.ProcessingContext
@@ -72,12 +72,8 @@ object MiseTomlPsiPatterns {
      * ```
      */
     val onTaskDependsArray =
-        StandardPatterns.or(
-            psiElement<TomlArray>()
-                .withParent(taskProperty("depends")),
-            psiElement<TomlArray>()
-                .withParent(taskProperty("depends_post")),
-        )
+            psiElement<TomlArray>().withParent(taskProperty("depends")) or
+            psiElement<TomlArray>().withParent(taskProperty("depends_post"))
     val inTaskDependsArray = miseTomlPsiElement<PsiElement>().inside(onTaskDependsArray)
 
     /**
@@ -94,15 +90,10 @@ object MiseTomlPsiPatterns {
      * ```
      */
     val onTaskDependsString =
-        StandardPatterns.or(
-            miseTomlStringLiteral()
-                .withParent(taskProperty("depends")),
-            miseTomlStringLiteral()
-                .withParent(taskProperty("depends_post")),
-        )
-    val inTaskDependsString = miseTomlPsiElement<PsiElement>().inside(onTaskDependsString)
+            miseTomlStringLiteral().withParent(taskProperty("depends")) or
+            miseTomlStringLiteral().withParent(taskProperty("depends_post"))
 
-    inline fun <reified I : PsiElement> psiElement() = PlatformPatterns.psiElement(I::class.java)
+    val inTaskDependsString = miseTomlPsiElement<PsiElement>().inside(onTaskDependsString)
 
     fun <T : Any, Self : ObjectPattern<T, Self>> ObjectPattern<T, Self>.with(
         name: String,
@@ -116,4 +107,11 @@ object MiseTomlPsiPatterns {
                 ): Boolean = cond(t, context)
             },
         )
+
+}
+
+inline fun <reified I : PsiElement> psiElement(): PsiElementPattern.Capture<I> = PlatformPatterns.psiElement(I::class.java)
+
+inline infix fun <reified I : PsiElement> ElementPattern<out I>.or(pattern: ElementPattern<out I>): PsiElementPattern.Capture<I> {
+    return psiElement<I>().andOr(this, pattern)
 }