+++ modules/core/src/main/kotlin/com/github/l34130/mise/core/lang/resolve/MiseTomlReferenceContributor.kt
@@ -1,14 +1,14 @@
 package com.github.l34130.mise.core.lang.resolve
 
 import com.github.l34130.mise.core.lang.psi.MiseTomlPsiPatterns
-import com.intellij.patterns.StandardPatterns
+import com.intellij.patterns.PlatformPatterns
 import com.intellij.psi.PsiReferenceContributor
 import com.intellij.psi.PsiReferenceRegistrar
 
 class MiseTomlReferenceContributor : PsiReferenceContributor() {
     override fun registerReferenceProviders(registrar: PsiReferenceRegistrar) {
         registrar.registerReferenceProvider(
-            StandardPatterns.or(
+            PlatformPatterns.or(
                 MiseTomlPsiPatterns.miseTomlStringLiteral().withParent(MiseTomlPsiPatterns.onTaskDependsArray),
                 MiseTomlPsiPatterns.onTaskDependsString,
             ),