+++ modules/products/goland/src/test/kotlin/com/github/l34130/mise/goland/GoLandTest.kt
@@ -1,0 +1,9 @@
+package com.github.l34130.mise.goland
+
+import com.intellij.testFramework.fixtures.BasePlatformTestCase
+
+class GoLandTest : BasePlatformTestCase() {
+    fun testGoLand() {
+        println("GoLandTest.testGoLand")
+    }
+}