+++ modules/products/gradle/src/test/kotlin/com/github/l34130/mise/gradle/GradleTest.kt
@@ -1,0 +1,9 @@
+package com.github.l34130.mise.gradle
+
+import com.intellij.testFramework.fixtures.BasePlatformTestCase
+
+class GradleTest : BasePlatformTestCase() {
+    fun testGradle() {
+        println("GradleTest.testGradle")
+    }
+}