+++ modules/products/idea/src/test/kotlin/com/github/l34130/mise/idea/IdeaTest.kt
@@ -1,0 +1,9 @@
+package com.github.l34130.mise.idea
+
+import com.intellij.testFramework.fixtures.BasePlatformTestCase
+
+class IdeaTest : BasePlatformTestCase() {
+    fun testIdea() {
+        println("IdeaTest.testIdea")
+    }
+}